Move tests into separate test file and fix neats
